Extract #answer_question method into module.
This method was repeated three times so has been extracted to
prevent repetition.
