+++ spec/features/brexit_checker/question_navigation_spec.rb
@@ -1,6 +1,8 @@
 require "spec_helper"
 
 RSpec.feature "Navigating Brexit Checker questions", type: :feature do
+  include BrexitCheckerHelper
+
   scenario "navigate to previous question" do
     when_i_visit_the_brexit_checker_flow
     and_i_answer_travel_questions
@@ -31,12 +33,5 @@ RSpec.feature "Navigating Brexit Checker questions", type: :feature do
 
   def and_my_previous_activities_choice_is_selected
     expect(page).to have_checked_field("Take your pet")
-  end
-
-  def answer_question(key, *options)
-    question = BrexitChecker::Question.find_by_key(key)
-    expect(page).to have_content(question.text)
-    options.each { |o| find_field(o).click }
-    click_on "Next"
   end
 end
