+++ spec/features/brexit_checker/question_results_spec.rb
@@ -1,6 +1,8 @@
 require "spec_helper"
 
 RSpec.feature "Brexit Checker workflow", type: :feature do
+  include BrexitCheckerHelper
+
   scenario "business questions" do
     when_i_visit_the_brexit_checker_flow
     and_i_answer_citizen_questions
@@ -95,12 +97,5 @@ RSpec.feature "Brexit Checker workflow", type: :feature do
     if action.guidance_link_text
       expect(page).to have_link(action.guidance_link_text, href: action.guidance_url)
     end
-  end
-
-  def answer_question(key, *options)
-    question = BrexitChecker::Question.find_by_key(key)
-    expect(page).to have_content(question.text)
-    options.each { |o| find_field(o).click }
-    click_on "Next"
   end
 end
