+++ spec/features/brexit_checker/travel_question_spec.rb
@@ -1,6 +1,8 @@
 require "spec_helper"
 
-RSpec.feature "Brexit Checker travel question", type: :feature do
+RSpec.feature "Filtering options based on criteria", type: :feature do
+  include BrexitCheckerHelper
+
   scenario "User's living question choice is not displayed as a travel question option" do
     when_i_visit_the_brexit_checker_flow
     and_i_choose_uk_for_the_living_question
@@ -15,18 +17,12 @@ RSpec.feature "Brexit Checker travel question", type: :feature do
     click_on "Next"
     answer_question("living", "UK")
     click_on "Next"
+    click_on "Next"
   end
 
   def then_i_should_not_see_uk_as_an_option_for_the_travel_question
     question = BrexitChecker::Question.find_by_key("travelling")
     expect(page).to have_content(question.text)
     expect(page).not_to have_content("To the UK")
-  end
-
-  def answer_question(key, *options)
-    question = BrexitChecker::Question.find_by_key(key)
-    expect(page).to have_content(question.text)
-    options.each { |o| find_field(o).click }
-    click_on "Next"
   end
 end
