+++ spec/support/brexit_checker_helper.rb
@@ -1,0 +1,8 @@
+module BrexitCheckerHelper
+  def answer_question(key, *options)
+    question = BrexitChecker::Question.find_by_key(key)
+    expect(page).to have_content(question.text)
+    options.each { |o| find_field(o).click }
+    click_on "Next"
+  end
+end
